ml-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
prophet==1.1.5
cmdstanpy==1.2.2
joblib==1.3.0
requests==2.31.0
openpyxl==3.1.2
fpdf2==2.7.0
schedule==1.2.0
python-dateutil==2.8.2
matplotlib==3.7.1
